+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/models/__init__.py
@@ -44,9 +44,17 @@ try:
     from ._models_py3 import ProxyResource
     from ._models_py3 import RegenerateTestKeyRequestPayload
     from ._models_py3 import Resource
+    from ._models_py3 import ResourceSku
+    from ._models_py3 import ResourceSkuCapabilities
+    from ._models_py3 import ResourceSkuLocationInfo
+    from ._models_py3 import ResourceSkuRestrictionInfo
+    from ._models_py3 import ResourceSkuRestrictions
+    from ._models_py3 import ResourceSkuZoneDetails
     from ._models_py3 import ResourceUploadDefinition
     from ._models_py3 import ServiceResource
     from ._models_py3 import ServiceSpecification
+    from ._models_py3 import Sku
+    from ._models_py3 import SkuCapacity
     from ._models_py3 import TemporaryDisk
     from ._models_py3 import TestKeys
     from ._models_py3 import TraceProperties
@@ -87,9 +95,17 @@ except (SyntaxError, ImportError):
     from ._models import ProxyResource
     from ._models import RegenerateTestKeyRequestPayload
     from ._models import Resource
+    from ._models import ResourceSku
+    from ._models import ResourceSkuCapabilities
+    from ._models import ResourceSkuLocationInfo
+    from ._models import ResourceSkuRestrictionInfo
+    from ._models import ResourceSkuRestrictions
+    from ._models import ResourceSkuZoneDetails
     from ._models import ResourceUploadDefinition
     from ._models import ServiceResource
     from ._models import ServiceSpecification
+    from ._models import Sku
+    from ._models import SkuCapacity
     from ._models import TemporaryDisk
     from ._models import TestKeys
     from ._models import TraceProperties
@@ -101,6 +117,7 @@ from ._paged_models import CertificateResourcePaged
 from ._paged_models import CustomDomainResourcePaged
 from ._paged_models import DeploymentResourcePaged
 from ._paged_models import OperationDetailPaged
+from ._paged_models import ResourceSkuPaged
 from ._paged_models import ServiceResourcePaged
 from ._app_platform_management_client_enums import (
     ProvisioningState,
@@ -113,6 +130,9 @@ from ._app_platform_management_client_enums import (
     RuntimeVersion,
     DeploymentResourceProvisioningState,
     DeploymentResourceStatus,
+    SkuScaleType,
+    ResourceSkuRestrictionsType,
+    ResourceSkuRestrictionsReasonCode,
 )
 
 __all__ = [
@@ -150,9 +170,17 @@ __all__ = [
     'ProxyResource',
     'RegenerateTestKeyRequestPayload',
     'Resource',
+    'ResourceSku',
+    'ResourceSkuCapabilities',
+    'ResourceSkuLocationInfo',
+    'ResourceSkuRestrictionInfo',
+    'ResourceSkuRestrictions',
+    'ResourceSkuZoneDetails',
     'ResourceUploadDefinition',
     'ServiceResource',
     'ServiceSpecification',
+    'Sku',
+    'SkuCapacity',
     'TemporaryDisk',
     'TestKeys',
     'TraceProperties',
@@ -165,6 +193,7 @@ __all__ = [
     'CustomDomainResourcePaged',
     'DeploymentResourcePaged',
     'OperationDetailPaged',
+    'ResourceSkuPaged',
     'ProvisioningState',
     'ConfigServerState',
     'TraceProxyState',
@@ -175,4 +204,7 @@ __all__ = [
     'RuntimeVersion',
     'DeploymentResourceProvisioningState',
     'DeploymentResourceStatus',
+    'SkuScaleType',
+    'ResourceSkuRestrictionsType',
+    'ResourceSkuRestrictionsReasonCode',
 ]