+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/models/_paged_models.py
@@ -103,3 +103,16 @@ class OperationDetailPaged(Paged):
     def __init__(self, *args, **kwargs):
 
         super(OperationDetailPaged, self).__init__(*args, **kwargs)
+class ResourceSkuPaged(Paged):
+    """
+    A paging container for iterating over a list of :class:`ResourceSku <azure.mgmt.appplatform.models.ResourceSku>` object
+    """
+
+    _attribute_map = {
+        'next_link': {'key': 'nextLink', 'type': 'str'},
+        'current_page': {'key': 'value', 'type': '[ResourceSku]'}
+    }
+
+    def __init__(self, *args, **kwargs):
+
+        super(ResourceSkuPaged, self).__init__(*args, **kwargs)